Move writeOverridesToWorkspace into the pnpm utils module
It's a generic workspace-manifest helper, not command-specific glue, and
pnpm.mjs already owns readPackageJson/writePackageJson. Tests move alongside
it into pnpm.test.mjs.
